+++ src/toy_crypto/sec_games.py
@@ -1,6 +1,6 @@
 from collections.abc import Callable
 import secrets
-from typing import Generic, TypeVar
+from typing import Generic, Optional, TypeVar
 
 K = TypeVar("K")
 """Unbounded type variable intended for any type of key."""
@@ -13,7 +13,7 @@ class IndCpa(Generic[K]):
         encryptor: Callable[[K, bytes], bytes],
     ) -> None:
         """
-        Meta setup for IND-CPA game.
+        IND-CPA game for symmetric encryption.
 
         Takes an encryptor, which is a function that takes a key and bytes
         and outputs bytes. And takes a
@@ -28,22 +28,44 @@ class IndCpa(Generic[K]):
 
         self._key_gen = key_gen
         self._encryptor = encryptor
+        self._recent_b: Optional[bool]
 
-    def new_challenger(
-        self,
-    ) -> Callable[[bytes, bytes], tuple[bytes, bool]]:
-        """Creates a new challenger with random key."""
+        self._key: Optional[K] = None
+        self._b: Optional[bool] = None
 
-        key = self._key_gen()
-        b = secrets.choice([True, False])
+    def initialize(self) -> None:
+        """Challenger picks key and a b."""
+        self._key = self._key_gen()
+        self._b = secrets.choice([True, False])
 
-        def challenger(m0: bytes, m1: bytes) -> tuple[bytes, bool]:
-            if len(m0) != len(m1):
-                raise ValueError("Message lengths must be equal")
+    def encrypt_one(self, m0: bytes, m1: bytes) -> bytes:
+        """Challenger encrypts m0 if b is False, else encrypts m1.
 
-            m = m1 if b else m0
+        :raise ValueError: if lengths of m0 and m1 are not equal.
+        :raises Exception: if challenge isn't initialized.
+        """
 
-            encrypted = self._encryptor(key, m)
-            return encrypted, b
+        if len(m0) != len(m1):
+            raise ValueError("Message lengths must be equal")
 
-        return challenger
+        if self._b is None or self._key is None:
+            raise Exception("Challenge is not be properly initialized")
+
+        m = m1 if self._b else m0
+
+        return self._encryptor(self._key, m)
+
+    def finalize(self, guess: bool) -> bool:
+        """
+        True iff guess is the same as b of previously created challenger.
+
+        Also resets the challenger, as for this game you cannot call with
+        same key, b pair more than once.
+        """
+
+        adv_wins = guess == self._b
+
+        self._b = None
+        self._key = None
+
+        return adv_wins